move `lib/index` into root
