+++ package.json
@@ -3,10 +3,9 @@
   "version": "1.1.0",
   "description": "🐟 A fast, small-but-mighty, familiar ~fish~ router",
   "repository": "lukeed/trouter",
-  "main": "lib/index.js",
   "license": "MIT",
   "files": [
-    "lib"
+    "index.js"
   ],
   "author": {
     "name": "Luke Edwards",
